refactor(netutils): extract shared backends for ifupdown and sslutil
Split the ifupdown and sslutil applet packages so each command lives in
its own file over a shared backend, with public behavior unchanged.

ifupdown (#771): move the interface-state backend (interfaces-file
parsing, hook planning/running, the shared ifup/ifdown driver, and the
production shell runner) into backend.go, and keep ifup, ifdown, and
ifplugd entrypoints in their own files.

sslutil (#772): move the shared TLS/session setup (cert/key loading,
tls.Config construction, ServeTLS/DialAndPipe connection plumbing) into
tls.go, and keep ssl_client and ssl_server entrypoints in their own
files. Add ServerConfig/ClientConfig helpers for the cert-load and
client-config setup, covered by focused tests.

Closes #771
Closes #772
